add edit functionality to the application

+++ controllers/playController.js
@@ -36,7 +36,7 @@ router.post('/create', isUser(), async(req, res) => {
     }
 });
 
-router.get('/details/:id', async(req, res) => {
+router.get('/details/:id', isUser(), async(req, res) => {
     try {
         const id = req.params.id;
 
@@ -52,9 +52,55 @@ router.get('/details/:id', async(req, res) => {
     };
 });
 
+router.get('/edit/:id', isUser(), async(req, res) => {
+    const id = req.params.id;
+    try {
+        const play = await req.storage.getPlayById(id);
+
+        if (play.author != req.user._id) {
+            throw new Error('Cannot edit play you havn\'t created');
+        };
+
+        res.render('play/edit', { play });
+    } catch (error) {
+        console.log(error.message);
+        res.redirect(`/play/details/${id}`);
+    };
+});
+
+router.post('/edit/:id', isUser(), async(req, res) => {
+    const id = req.params.id;
+    try {
+        const play = await req.storage.getPlayById(id);
+        if (play.author != req.user._id) {
+            throw new Error('Cannot edit play you havn\'t created');
+        };
+
+        await req.storage.editPlay(id, req.body);
+        res.redirect('/');
+    } catch (error) {
+        const ctx = {
+            errors: parseError(error),
+            play: {
+                _id: id,
+                title: req.body.title,
+                description: req.body.description,
+                imageUrl: req.body.imageUrl,
+                public: Boolean(req.body.public)
+            }
+        };
+        res.render('play/edit/', ctx);
+    }
+});
+
 router.get('/delete/:id', async(req, res) => {
     let id = req.params.id;
     try {
+        const play = await req.storage.getPlayById(id);
+
+        if (play.author != req.user._id) {
+            throw new Error('Cannot delete play you havn\'t created');
+        };
         await req.storage.deletePlay(id);
         res.redirect('/');
     } catch (error) {
@@ -62,5 +108,7 @@ router.get('/delete/:id', async(req, res) => {
         res.redirect(`/play/details/${id}`);
     };
 });
+
+
 
 module.exports = router;
